requirements.txt
fastapi
uvicorn
openai
chromadb
pypdf
sentence-transformers
stix2
python-dotenv
ddgs
python-multipart
sse-starlette
